+++ creme/preprocessing/__init__.py
@@ -1,7 +1,6 @@
 """
 Module for preparing data to make it compatible with certain machine learning algorithms.
 """
-from .discard import Discarder
 from .feature_hasher import FeatureHasher
 from .one_hot import OneHotEncoder
 from .poly import PolynomialExtender
@@ -10,7 +9,6 @@ from .scale import MinMaxScaler
 
 
 __all__ = [
-    'Discarder',
     'FeatureHasher',
     'MinMaxScaler',
     'OneHotEncoder',